app/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pyyaml>=6.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.1

# LLM 翻譯支援
openai>=1.12.0
google-generativeai>=0.3.0
